EntityInsertPanel: move to @labkey/components/entities
